Update the README.md file
Raphael corrected mistakes such as ...
The C of Controller is ALWAYS capitalized when referring to the Avi Controller.
There has never been anything named Netscaler. It is NetScaler.
Avi has coined two special words (at least): DataScript and ControlScript. Under NO circumstances do we right datascript, Datascript, data script, controlscript, Controlscript, or control script. What we do in the first week of our employment is familiarize ourselves with the Glossary. 
Under NO circumstances do we write AVI. Proper usage is likewise documented in the Glossary. If and only if we are referring to the Avi object do we write VirtualService. Otherwise it is always virtual service or VS.